renew_migration_script/requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0 